added some documentation
