+++ app/safe/internal/state/internal.go
@@ -1,0 +1,26 @@
+/*
+ * .-'_.---._'-.
+ * ||####|(__)||   Protect your secrets, protect your business.
+ *   \\()|##//       Secure your sensitive data with Aegis.
+ *    \\ |#//                    <aegis.ist>
+ *     .\_/.
+ */
+
+package state
+
+import "encoding/json"
+
+const selfName = "aegis-safe"
+
+type AegisInternalCommand struct {
+	LogLevel int `json:"logLevel"`
+}
+
+func evaluate(data string) *AegisInternalCommand {
+	var command AegisInternalCommand
+	err := json.Unmarshal([]byte(data), &command)
+	if err != nil {
+		return nil
+	}
+	return &command
+}
