+++ app/sentinel/internal/safe/post.go
@@ -95,8 +95,8 @@ func Post(workloadId, secret, namespace, backingStore string, useKubernetes bool
 			bs = data.File
 		case data.Memory:
 			bs = data.Memory
-		case data.Cluster:
-			bs = data.Cluster
+		default:
+			bs = data.Memory
 		}
 	}
 
@@ -106,6 +106,8 @@ func Post(workloadId, secret, namespace, backingStore string, useKubernetes bool
 		f = data.Json
 	case data.Yaml:
 		f = data.Yaml
+	default:
+		f = data.None
 	}
 
 	sr := reqres.SecretUpsertRequest{
